+++ packages/camera/camera_android_camerax/android/src/main/java/io/flutter/plugins/camerax/ImageCaptureHostApiImpl.java
@@ -87,6 +87,10 @@ public class ImageCaptureHostApiImpl implements ImageCaptureHostApi {
   @Override
   public void takePicture(
       @NonNull Long identifier, @NonNull GeneratedCameraXLibrary.Result<String> result) {
+    if (context == null) {
+      throw new IllegalStateException("Context must be set to take picture.");
+    }
+
     ImageCapture imageCapture = getImageCaptureInstance(identifier);
     final File outputDir = context.getCacheDir();
     File temporaryCaptureFile;